Remove separate sync version
