+++ src/realm/sync/dump_command.cpp
@@ -15,7 +15,7 @@
 #include <realm/util/load_file.hpp>
 #include <realm/list.hpp>
 #include <realm/group.hpp>
-#include <realm/sync/version.hpp>
+#include <realm/version.hpp>
 
 using namespace realm;
 
@@ -455,7 +455,7 @@ int main(int argc, char* argv[])
 #else
             build_mode = "Release";
 #endif
-            std::cerr << "RealmSync/" REALM_SYNC_VER_STRING " (build_mode=" << build_mode << ")\n";
+            std::cerr << "RealmSync/" REALM_VERSION_STRING " (build_mode=" << build_mode << ")\n";
             return EXIT_SUCCESS;
         }
