+++ src/realm/sync/server_index_command.cpp
@@ -10,7 +10,7 @@
 #include <realm/util/load_file.hpp>
 #include <realm/sync/noinst/server_history.hpp>
 #include <realm/db.hpp>
-#include <realm/sync/version.hpp>
+#include <realm/version.hpp>
 
 using namespace realm;
 
@@ -214,7 +214,7 @@ int main(int argc, char* argv[])
 #else
             build_mode = "Release";
 #endif
-            std::cerr << "RealmSync/" REALM_SYNC_VER_STRING " (build_mode=" << build_mode << ")\n";
+            std::cerr << "RealmSync/" REALM_VERSION_STRING " (build_mode=" << build_mode << ")\n";
             return EXIT_SUCCESS;
         }
