+++ src/realm/sync/stat_command.cpp
@@ -14,7 +14,7 @@
 #include <realm/group.hpp>
 #include <realm/replication.hpp>
 #include <realm/bplustree.hpp>
-#include <realm/sync/version.hpp>
+#include <realm/version.hpp>
 #include <realm/sync/protocol.hpp>
 
 using namespace realm;
@@ -787,7 +787,7 @@ int main(int argc, char* argv[])
 #else
             build_mode = "Release";
 #endif
-            std::cerr << "RealmSync/" REALM_SYNC_VER_STRING " (build_mode=" << build_mode << ")\n";
+            std::cerr << "RealmSync/" REALM_VERSION_STRING " (build_mode=" << build_mode << ")\n";
             return EXIT_SUCCESS;
         }
