+++ src/realm/sync/verify_server_file_command.cpp
@@ -8,7 +8,7 @@
 #include <realm/util/load_file.hpp>
 #include <realm/sync/noinst/server_history.hpp>
 #include <realm/db.hpp>
-#include <realm/sync/version.hpp>
+#include <realm/version.hpp>
 
 using namespace realm;
 
@@ -89,7 +89,7 @@ int main(int argc, char* argv[])
 #else
             build_mode = "Release";
 #endif
-            std::cerr << "RealmSync/" REALM_SYNC_VER_STRING " (build_mode=" << build_mode << ")\n";
+            std::cerr << "RealmSync/" REALM_VERSION_STRING " (build_mode=" << build_mode << ")\n";
             return EXIT_SUCCESS;
         }
