+++ test/client/main.cpp
@@ -45,7 +45,7 @@
 #include <realm/util/base64.hpp>
 #include <realm/util/json_parser.hpp>
 #include <realm/sync/impl/clock.hpp>
-#include <realm/sync/version.hpp>
+#include <realm/version.hpp>
 #include <realm/sync/protocol.hpp>
 #include <realm/sync/client.hpp>
 
@@ -1107,7 +1107,7 @@ int main(int argc, char* argv[])
     sync_base_logger.set_level_threshold(sync_log_level);
     util::PrefixLogger sync_logger{"Sync: ", sync_base_logger};
     sync::Client::Config config;
-    config.user_agent_application_info = "TestClient/" REALM_SYNC_VER_STRING;
+    config.user_agent_application_info = "TestClient/" REALM_VERSION_STRING;
     config.logger = &sync_logger;
     config.one_connection_per_session = connection_per_session;
     config.dry_run = dry_run;
